init: Continue on a modules-directory mismatch
During kernel development it is very useful to boot a kernel from tftp
using the existing on-device root disk.

The modified kernel typically has a different uname, e.g. with a -dirty
suffix or a different commit hash.

At present this causes static_node_tool to produce a fatal error. The
boot then silently hangs, with no message present even if the serial
console is enabled.

This is very, very confusing. It can cause lots of bisecting to look
for a boot failure that has nothing to do with the code itself.

This should not happen with a consistent build, so it may be safe to
change the error to be non-fatal, so that the boot can proceed, even if
some features are missing.

BUG=b:316592048
TEST=build and boot on rockpro64
See output:
   ERROR static_node_tool: [static_node_tool.cc(42)] Could not read in
     list of modules (continuing anyway): No such file or directory (2)

Change-Id: I2838ea224910df1b4fbe970f91a03708df0b8996
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5122397
Tested-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>