cryptohome: Add a interface for sending signals
This creates a new standard interface for sending signals, and converts
one of the existing callback-based interfaces over to using it. The
intent would be to convert the other existing signals as well, but this
initial version just demonstrates the usage and provides an easier place
to add new signals. Conversion of the other existing signals can be done
more lazily.

The benefit of this approach over the existing callback one is that it
reduces boilerplate somewhat (and the boilerplate you do need to add for
every new signal is much more trivial). It also make it easier to add
more signals to existing layers of the code that already send at least
one, because you would no longer need to forward more and more callbacks
to those other layers.

The main disadvantage is that you lose the more granular interface that
a single callback provides; if you expose the SignallingInterface to a
piece of code then it can send all signals, not just one. In practice I
believe this is more of a hindrance than a benefit.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I5a5f12ceaef8e95f165e2a7b39ad276ea04d7e70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5126914
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>