hiberman: Improve logging in resume()
The ResumeConductor is a bit weird in the sense that it may
resume from hibernate (if a resume is pending), but most of the
time it just sets things up for a possible future hibernation.
Currently the ResumeConductor emits a log "Beginning resume"
when it starts, which is misleading. Change this to two
different logs, one log or the other is emitted, depending on
whether a resume from hibernate is pending or not.

BUG=none
TEST=emerge-${BOARD} hiberman

Change-Id: I08e8b4b8207d10ae5a2dc5270be87484e7398ad3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5124119
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>