shill: portal: complete trial on portal-suspected results
This patch optimizes the time to show the captive portal sign-in UX
notification when the captive portal is detected as a "portal-suspected"
case. Similarly to the "portal-redirect" state, the result of the HTTPS
probe is not blocking the transition of the Network to the
"portal-suspected" state, and the sign-in UX can trigger as soon as the
response to the HTTP probe is received.

Also makes sure that the HTTP probe URL is correctly set for the
"portal-suspected" case corresponding to an invalid redirect code.

BUG=b:309175584
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I1e72c0eb3e0a8248664d633801e2e55af1aee9aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5137602
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>