featured: Support for dynamic vcpu boost experiment
Dynamic vcpu boosting is a new feature that boosts the priority of vcpu
tasks when the VM is running a latency sensitive workload on the vcpu.
More details here: go/vcpu_boosting_kvm

This CL enables the support for testing this feature in finch. Adds the
experiment in featured configuration and updates selinux policies for
allowing featured access to the kernel knob to enable this feature.

BUG=b:306637516
TEST=tested on local DUT and verified featured can enable the feature.

Change-Id: Ib3c8f2ffa03876eb1d5e505af5b92bd3f68a5807
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5136780
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Joel Fernandes <joelaf@google.com>
Commit-Queue: Joel Fernandes <joelaf@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: Vineeth Pillai <vineethrp@google.com>