fusebox: fix "mv foo already_exists"
Prior to this commit, when renaming a source file to a destination that
already existed, (1) the underlying move in the fusebox D-Bus server in
ash Chrome can succeed but (2) the intermediate move in the fusebox
daemon (the D-Bus client) would always fail.

For (1), see:
https://source.chromium.org/chromium/chromium/src/+/main:storage/browser/file_system/async_file_util.h;l=245;drc=aa2cf6a4fdd6ed15a3c4dfd3c9940618e78d5525

Here's an example of getting into this inconsistent state, starting in a
fusebox-served directory:

$ echo -n abc   > 3.txt
$ echo -n lmnop > 5.txt
$ ls -i -l ?.txt   # -i prints inode numbers, -l prints file sizes.
$ mv 3.txt 5.txt
$ ls -i -l ?.txt   # -l needs a stat; fusebox incorrectly says ENOENT
ls: cannot access '5.txt': No such file or directory

After the commit, the second "ls -i -l ?.txt" correctly lists "5.txt",
with a file size of 3 bytes.

BUG=b:299037589
TEST=manual

Change-Id: I52f593a5007c62a2c4bcddae4d0cbaf305c5ca25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5138580
Tested-by: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Nigel Tao <nigeltao@chromium.org>