cryptohome: Get rid of AuthSession::auth_factor_map()
Now that the auth factor maps are shared between all sessions, and
available via AuthFactorManager, it's not really necessary to have an
accessor for them exposed by AuthSession. Pretty much everything that
needs direct access to the map either already has access to the manager
or can very easily have it.

It does make some code a bit more verbose, because doing
GetAuthFactorMap requires passing in the username whereas with the
AuthSession version this was basically implied by the session, but I
don't think that by itself really justifies having this be a part of the
AuthSession interface.

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I3e08c47cbff8332a5a6432a3ed013e5225d94bfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5141386
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>