fbpreprocessor: wait for the policy to be available
When a user logs in for the first time there is a delay before the
policy is available after login. Add a 2s delay on login before
retrieving the policy.

BUG=b:317161262
TEST=FEATURES=test emerge-${BOARD} chromeos-base/fbpreprocessord
TEST=cros deploy $DUT chromeos-base/fbpreprocessord
TEST=restart fbpreprocessord
TEST=unenroll the device to go back to OOBE and add a managedchrome.com
user account. Observe in the fbpreprocessord log that the policy has
been successfully parsed on first login

Change-Id: I3afa5dcc4b7642f9e5c3a65a13a85942e9bfffce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5142554
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>