lorgnette: Save known devices across sessions
lorgnette is supposed to return the same deviceUuid each time it sees
the same device.  This works correctly when lorgnette remains running
between two discovery sessions.  However, lorgnete exits fairly quickly
after inactivity, so this state is frequently discarded.

In order to keep the deviceUuid across sessions, save the list of known
devices whenever enumeration of local devices is complete.  This allows
two improvements:

* When a client opens a scanner from a previously-issued connection
  string, other clients are not blocked from discovering that scanner
  even if lorgnette has exited in between.
* DeviceUuid values for the same device remain consistent across
  discovery sessions even if lorgnette exits in between.

To avoid building up an infinitely growing list of previously seen
devices, the cache is written to /run/lorgnette.  This directory is
cleared on each reboot, which will cause subsequent discovery sessions
to get a whole new set of IDs.  The scanner IDs provided by the public
getScannerList APIs that wrap discovery are explicitly documented as not
being reusable across login sessions, so this should not be a problem.

BUG=b:315831309, b:311196232
TEST=lorgnette_cli discover --show_details

Change-Id: Id29a3354f0ed4bb5d259a68cbab5033c6e24bcec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5146991
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>