shill: cellular: don't ignore abort on error while disconnecting tethering
The AbortTetheringOperation() method is in charge of cancelling the
current ongoing tethering operation and launching whatever other
operation is necessary to recover the original state. E.g. if a
multiplexed DUN connect is aborted, it should complete the original
operation with an error and right away launch a multiplexed DUN
disconnect operation. Cancelling a user requested tethering
disconnect is not a valid use case supported in this context.

The AbortTetheringOperation() method is also used to recover from
errors during any of the connect/disconnect tethering operations. In
this case, though, we must make sure we complete all types of
operations, including the disconnect ones.

The check of whether the disconnection abort should be processed or
not is therefore moved out of AbortTetheringOperation() in the
ReleaseTetheringNetwork() method, which is the one processing
operation abort requests triggered by the user.

BUG=b:316013990
TEST=FEATURES="test" emerge-$BOARD chromeos-base/shill

Change-Id: Icfeca874cb1d7f5d5d11f08858073d7b63e3b082
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5148468
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>