shill: cellular: fix state on DUN as DEFAULT disconnection failure
A failure in a DUN as DEFAULT disconnection procedure means that the
attempt to connect back to the DEFAULT APN failed, so we must report
the modem as disconnected, or otherwise it will be left in "Connected"
or "Linked" state, and reconnection won't be possible.

In order to handle this properly, we must make sure that
AbortTetheringOperation() unconditionally completes the current
tethering operation, regardless of whether it is a disconnection
or connection attempt being aborted.

BUG=b:316013990
TEST=FEATURES="test" emerge-$BOARD chromeos-base/shill

Change-Id: Iaa50bc7c3a1860a8e139bc43522e12b458349e62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5148469
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>