cryptohome: sepolicy: Add key_store_certs path
A previous CL added the directory to store recoverable key store
certificates under .shadow/, but forgot to add the selinux file
contexts. Add the file context and rules, as well as rename the path to
key_store_certs to be shorter. The renaming is fine because the path
hasn't been used and won't be created on real devices yet.

BUG=b:309734008
TEST=tast run $DUT security.SELinuxFiles*

Cq-Depend: chromium:5149449
Change-Id: I231fddfc5ba1ce4e9921fa73678c70343672d5f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5148394
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Qijiang Yūki Ishii <fqj@google.com>