cryptohome: Hardcode the current certificate list
To allow cryptohome to start generating recoverable key stores, hardcode
the current certificate list. We have a CL that implements actual
certificate fetching already, but we want to have more discussion around
where is the best place to send the network requests (probably not in
cryptohomed). This CL will enable the passkey feature to launch in the
meantime. In addition, we might want to hardcode a default list version
even if we have the fetcher, to be more robust against network failures.

In addition, mock the cert provider in fuzz test and unit tests because
it might cause unnecessary latency or some failures that the test isn't
intended to test.

BUG=b:309734008
TEST=cryptohome unittests

Change-Id: I45c3a55a17c1f30df2991b5711cf941afc9e270a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5151753
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>