cryptohome: Generate key store for password
Generate recoverable key stores for password factor too. Note that we
don't want to use GAIA passwords for passkey recovery, but it's not
possible to control it from cryptohome because we can't distinguish.
It's up to Chrome to only send the hash information for local passwords
such that we don't generate recoverable key stores for GAIA passwords.

BUG=b:300578783
TEST=cryptohome unittests

Change-Id: I9920fa3156072547ada9db3e6fb38d6f62e314ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5149393
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>