system-proxy: Duplicate CURL socket fd from CURL handle
We need to create a patchpanel::Socket instance from CURL handle.
However, the file descriptor is owned by the CURL handle. If we use the
file descriptor at patchpanel::Socket directly, the file descriptor will
be closed twice. To workaround this issue, currently we release the
base::ScopedFD at the destructor of CurlSocket.

This CL duplicates the file descriptor of the socket, and passes the
duplicated one to patchpanel::Socket.

BUG=b:317751088
BUG=crbug:1070732
TEST=FEATURES=test emerge-brya system-proxy

Change-Id: I1ac849cd8155fb256fbd1552d4e929627ac18496
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5152396
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>