hiberman: Refactor hibernate_inner() to avoid extra mount
Delineate the control flow between the return of the
hibernate_inner() procedure and it's enclosing hibernate()
outer procedure in the combinatorics of when a resume fails
or not and how to proceed.

The aim here is for the CFG to read more linearly and thus
reduce down the number of meta volume mount/umount cycles by
one.

BUG=b:309654982
TEST=`emerge-nissa hiberman`.

Change-Id: Ib8753ea62095843bf8533cb4da93f28d3c1a1864
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5151741
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>