hiberman: Take phantom ref to mount for TimestampFile
The TimestampFile methods expect that a hibermeta
volume is ambiently mounted however there is not type
enforcement of this. Therefore have the type constructor
consume a phantom reference to bind its pre-requirements into
its type. This ensures member methods cannot be dispatched
without an ActiveMount being in scope.

This is akin to the LogFile construct and so later shall
be consolidated into the same namespace.

BUG=b:309654982
TEST=`emerge-nissa hiberman`.

Change-Id: Ie8962209668e4a8339059ebaebd23c24f1e3d517
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5151740
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>