hiberman: Exit resume_init with magic value when no resume from hibernate is pending
resume_init() currently returns a CookieError to main when no
resume from hibernate is pending. This results in an error
message which can be confusing. Instead exit directly when no
resume is pending. Use a special exit code to allow
chromeos_startup to differentiate between 'no resume pending'
and 'some kind of error occcurred'. With that chromeos_startup
can omit the warning it normally logs when hiberman exits with
a non-exit code.

Also update the log message of the no resume case, say that
no resume is pending, instead of talking about the hibernate
cookie, which is an implementation detail.

BUG=b:318006759
TEST=boot brya
  => no cookie error message or warning from chromeos_startup
     about hiberman

Change-Id: I94ac6162f8df30e47fe369b28e12ed8ffde3f3ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5155176
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>