cryptohome: Add timestamp field to key store
Add a timestamp field to the recoverable key store, so that it can
become one of the factors for Chrome to choose the most preferred key
store to upload. It can also be used for popping notification that
prompts user to enter the knowledge factors again once in a while to
prevent their recovery factor becoming stale.

BUG=b:312628857
TEST=build ok

Change-Id: I95aa1520b852b21dfd474b27d7f722fbf93d1011
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5154652
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>