shill: cellular: don't bring IPA net down on disable
Bringing down the network interface during the cellular technology
disable has some unexpected consequences.

If the modem radio is powered up very quickly after the IPA network
interface is brought down (e.g. upon a quick suspend/resume
operation), the modem enters a bad state in which it seems to not
correctly receive SLAAC RA from the network during an IPv6
attach (which leads to the modem being kicked out the T-Mobile network
as it ends up retrying to attach with IPv4-only), and also makes the
IPA hardware get stuck during system shutdown.

Avoid all these issues by making sure the network interface is never
brought down by shill.

This issue was introduced during the multiplexing support for trogdor,
but not because we now use MM's own multiplexing, the reason for this
problem is that the shill Cellular device is now associated to the
rmnet_ipa0 interface. When using the qc-netmgr, the shill Cellular
device was associated to the rmnet_data0 interface, and there was no
problem bringing that interface down.

BUG=b:305930007,b:302714371
TEST=On trogdor+T-Mobile
  $ tast --verbose run localhost:2222 cellular.ShillSafetyDance

Change-Id: Ic3a5848fd57f045f9007a80cb11f184943fadfb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5155970
Reviewed-by: Pavan Holla <pholla@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>