ml: Add support for langid events and multilingual configuration in soda.
Add support to the soda mojo for configuration to allow for
multilingual soda on chromeos. This adds support to the extended soda
config as per the rest of chrome configurations.

Add support to return LANGID events from chromeos in this case, and
add a unit test for the new response type in the response union.

BUG=b:257166700
TEST=compile

Change-Id: Iace5929d2745d097fdc80602e0ce1628de314e6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5149244
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: Evan Liu <evliu@google.com>
Tested-by: Rob Schonberger <robsc@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>