cryptohome: Improve session handling in Remove
This makes the session handling somewhat more robust. First, it
terminates all sessions for the user, not just the one being used for
the Remove. Second, it more properly manages the "in-use" lifetime of
the session, ensuring it's kept alive until the actual homedir removal
completes and then explicitly terminating it after that.

In practice on a real system this shouldn't make a big difference as
this generally only happens along with an unmount which terminates all
sessions anyway, but cryptohomed shouldn't assume a specific user flow.

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I8404b8a29645298dce55c46ad4b180219021f02d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5160655
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>