fusebox: implement Flush on the client
Note that for the concept of "persist this to disk (or other permanent
storage)", FUSE calls this "fsync" while Chrome's //storage C++ code
calls this "flush". FUSE has a separate concept called "flush", which is
largely unrelated.

Prior to this commit, fusebox did not do anything explicitly about
fsync. In terms of the FUSE protocol (between the kernel and the FUSE
server), libfuse falls back to returning ENOSYS (not implemented):
https://github.com/libfuse/libfuse/blob/3c7ba57050ae84679a81dde6bbce876826c25420/lib/fuse_lowlevel.c#L1497
The kernel then translates ENOSYS to 0 (no error):
https://github.com/torvalds/linux/blob/ab27740f76654ed58dd32ac0ba0031c18a6dea3b/fs/fuse/file.c#L599

Prior to this commit, fsync requests from the kernel to the FUSE server
(the fusebox daemon) were no-ops (but reported onwards as no error).
After this commit, fsync requests are forwarded on from the fusebox
daemon to Chrome and its storage::FileStreamWriter::Flush error code
(success or failure) is passed back along the chain.

This is one half of a multi-repo change. The other half is at
https://chromium-review.googlesource.com/c/chromium/src/+/5167491

BUG=b:255703917
TEST=manual

Change-Id: I7596ecfebefce3524c60bdd3e40de6b875bb7dfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5167789
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Tested-by: Nigel Tao <nigeltao@chromium.org>