cryptohome: Use SignallingInterface for fingerprint scans
This converts the AuthScanResult signal over to using the new interface.
It's a bit more involved than the other signals that were initially
converted since it has to get pushed down to the fingerprint service but
overall this seems like a simplification as it replaces a chain of
callback registration that cuts through three layers of objects.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I16c7338c72a2f167585c630c36742794da22ad6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5170626
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>