shill: cellular: Add obsolence warning for crosh set_apn
The crosh command set_apn is very outdated and seems to be unused.
The command is no longer needed since the UI can do the same thing.
Add a obsolence warning message to the script so users can report if the
script is still being used.

BUG=b:317421269
TEST=deploy shill to limozeen, run the script and verify new message

Change-Id: Ic25862dc7b8e49a756c6c9d76192d49137e63415
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5182222
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>