vm_tools: vsh: improve closed stream handling
If the guest sends the host a zero-length DataMessage, the vsh client
will close the file descriptor corresponding to that stream (either
stdout or stderr). In that case, we previously continued on to execute
the code that writes the received data to the file descriptor, even if
it had just been closed. This seems to not be a real problem in practice
since base::WriteFileDescriptor() will not actually call write() if
zero-length data is provided, but we should not be passing an invalid,
closed file descriptor to WriteFileDescriptor(), so add a return
statement to exit early in the zero-length buffer case.

Any further DataMessages targeting a closed stream would indicate a bug
in the server; this should not happen, but to be more defensive (and to
provide a better error if it does happen), the target file descriptor's
validity is now checked before attempting to write to it in the client.

BUG=b:319174128
TEST=tast run $DUT crostini.InputLatency.bookworm

Change-Id: I16e08c6d682fe5e4d131265b9d3a8afd13e0080d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5182030
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Ryuichiro Chiba <chibar@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>