shill: Workaround for wpa_supplicant issue.
wpa_supplicant's SignalPoll wraps result dictionary by variant
unexpectedly. Current code unexpectedly works with the case
because it automatically unwraps variant regardless of D-Bus signature.

That code is going to be removed to make it more strict, but that
will cause an issue on SignalPoll. This CL introduces the workaround
so that even when we remove that code, shill should keep working.
This workaround should be short term and be reverted when upstream is
fixed and then it is rolled.

BUG=b:318430162
TEST=Tryjob. Manually run DUT with wifi, and no crash is found.

Change-Id: I2dfd4964bedb7e8e65e8f99cd5ac220e48b363ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5185156
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>