vm_tools: concierge: enable battery device for TerminaVm
This adds the Goldfish battery emulation for all guests that derive from
TerminaVm to give the guest information about the current power state
and battery level.

There should be no observable effect with this change alone since the
CONFIG_GOLDFISH_BATTERY kernel option is not enabled yet for any of the
relevant guest kernels, but the presence of the battery device can be
verified by looking for the GFSH0001 entry in ACPI on x86 systems.

BUG=b:319475173
TEST=Deploy and start Crostini on x86 and arm

Change-Id: I52ede63212c0af0015471d1bf50aecbc6339b4cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5187406
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>