system_api: Add OverrideState field to FeatureOverride
Add an OverrideState enum to indicate if a feature's override state is
enabled, disabled, or to use the default. This will eventually replace
the enabled field, which is now marked as deprecated so updates can be
upreved into Chrome without breaking existing code.

Using the OverrideState enum instead of a bool allows us to represent
three states instead of two, which is what Chrome does.

A follow-up CL will remove all existing uses of the enabled field.

BUG=b:319686310
TEST=CQ

Change-Id: Idbc488608819f022a2b2a21d2d548b93aa70ed54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5190173
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Kendrake Tsui <kendraketsui@google.com>
Tested-by: Kendrake Tsui <kendraketsui@google.com>