arc-setup: Remove stale MediaProvider databases on container R
http://ag/25870771 skips the MediaProvider database migrtion on ARC R.
However, for container R devices that already went through or in the
middle of the migration, the MediaProvider databases should be removed,
because otherwise ModernMediaScanner throws an exception every time it
scans a file that has already been scanned by LegacyMediaProvider (see
b/179233339).

Also, since we do not perform the migration, the LegacyMediaProvider
databases are just a waste of disk space, and should be removed too.

This CL adds a logic in arc-setup's OnBootContinue() to 1) remove the
MediaProvider databases just once (if they exist), and 2) remove the
LegacyMediaProvider databases.

BUG=b:319460942
TEST=Manually confirm that:
TEST=1) MediaProvider databases are removed and .media... is created
TEST=   when the databases exist and .media... does not exist (first
TEST=   regular boot).
TEST=2) Nothing happens when .media... exists (second regular boot).
TEST=3) .media... is created when databases do not exist and .media...
TEST=   does not exist (P-to-R upgrade and new opt-in).
TEST=4) LegacyMediaProvider databases are removed if they exist.

Change-Id: I3e0c27e5efd55bebf778bc4fb7ff7c75a212a480
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5196415
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>