shill: portal: Network::EventHandler notifies start errors
This patch changes Network::EventHandler::OnNetworkValidationStart and
Network::EventHandler::OnNetworkValidationStop to communicate to
listeners if network validation started normally or failed to start,
stopped normally or was stopped after a failure.

If network validation fails to start or restart, the Network was not
configured properly and the Network should be considered as a network
without Internet access.

This allows Device and Service to not have to explicitly update the
Service state when Network::StartPortalDetection() fails.

BUG=b:265607618
BUG=b:305129516
BUG=b:314693271
TEST=FEATURES=test emerge-$BOARD shill. Manual tests: connected WiFi and
used Manager.RecheckPortal, Manager.CheckPortalList, Service proxy
configuration in the UI, and iptables DROP rules for blocking HTTPS
return traffic to exercises Service::UpdatePortalDetector,
Network::EventHandler callbacks, and the PortalDetector state machine.
Verified with office captive portal that portal detection works.

Change-Id: I4240d458fc446d9612450300d093881d83bac0a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5199746
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>