shill: avoid Network access null NetworkMonitor
Inside the Network class, the instances |network_monitor_|,
|dhcp_controller_|, and |slaac_controller_| are created at Start() and
destroyed at StopInternal(). That means these instances exist if and
only if the state is not kIdle. We could use this restriction to avoid
accessing the |network_monitor_| when it's null.

This CL fixes the Network::OnSetupConnectionFinished() and
Network::Start() to make sure the restriction mentioned above always
valid.

BUG=b:320165022
TEST=FEATURES=test emerge-brya shill

Change-Id: I8fca82955d9b8a6b3ffc04c9e2821f66f35c8944
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5197450
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>