resourced: Report shmem swap for crosvm
Shmem that has been swapped is not included in the SwapPss of any
process's smaps_rollup. In the kernel, PSS and shmem are abstractions in
different layers (address space and file systems), which makes resolving
this issue in the kernel very difficult.

On ChromeOS, the lack of shmem in SwapPss is primarily a problem because
crosvm uses shmem for guest memory. Rather than trying to solve the
general case in the kernel, we can deal with this specific problem by
parsing the smaps of any root crosvm processes and summing up the Swap
field of any memfd regions.

BUG=b:319364562
TEST=cargo test

Change-Id: I9257c3e5ececbc9db10a43705d200b3dd40d7228
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5197946
Commit-Queue: David Stevens <stevensd@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Shin Kawamura <kawasin@google.com>