net-base: Set IFA_ADDRESS in RTNLHandler::AddressRequest
Set both IFA_ADDRESS and IFA_LOCAL to the local interface address when
creating an RTNLMessage for modifying an address. By definition, they
should be the same when the interface is not used as an point-to-point
one.

Previously we only set IFA_LOCAL, and seems that the kernel can handle
it properly. Change it to follow the convention and avoid ERROR logs
in shill.

BUG=b:319342418
TEST=Manually switched wifi network several times, no longer observed
  the error log mentioned in $BUG.

Change-Id: I2567f1c73f1ea4b01a1e6b6d30686d50c5c1f148
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5196279
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>