attestation: Add recoverable key store interface
We want to put the recoverable key store service that require network
service (like fetching the certs from a static site) in an isolated
daemon. pca_agentd is determined to be the suitable daemon as it's used
to serve similar purposes (the network access part needed by
attestationd).

Add another dbus interface to the PcaAgent service, that allows the
fetched certificates to be queried, and also sending signal updates when
they're fetched.

BUG=b:320220928
TEST=build ok

Change-Id: I06ad7820cb164d37c7bed2293e85b9dea38b9c4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5195474
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>