shill: fix Passpoint BSS count
Since a BSS IE might be updated after it was discovered by a scan, the
count of HS2.0 compatible endpoint is not correct anymore and may raise
the CHECK() in BSSRemoved.

Instead of keeping track of the number of Passpoint compatible endpoint
and trigger an interworking selection later, we now check the endpoint
list directly to check if an interworking selection is needed.

BUG=b:316796681
TEST=tast run ${dut} network.PasspointSelection.* network.PasspointRoaming.*;
  FEATURES="test" emerge-$BOARD shill

Change-Id: Id0ebb653216a143a05afb01e249d25a05c6b53b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5200489
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Damien Dejean <damiendejean@google.com>
Tested-by: Damien Dejean <damiendejean@google.com>