shill: detect HS20 support change on BSS update
Hotspot 2.0 support (Passpoint) might be announced after a scan. In this
case, the change is discovered on a BSS properties update. The event is
received by the WiFiEndpoint and transmitted to the WiFiDevice. The
device now detects changes in HS20 support and schedules an interworking
selection if needed.

BUG=b:316796681
TEST=FEATURES="test" emerge-$BOARD shill
  tast run ${DUT} network.PasspointSelection.* network.PasspointRoaming.*

Change-Id: I240205eeefb3cb644b574d5d72e59e46026c39e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5211090
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>