shill: Fix off-by-one error in wifi_phy
wifi_phy uses the nl80211_iftype definitons from nl80211.h [1] to define
the list of valid wireless interface types. This enum defines a value
NL80211_IFTYPE_MAX which represents the largest valid value within the
iftype enum. When wifi_phy parses an interface type it compares it
agaist this value to ensure that the type is valid. It does this
incorrectly by treating NL80211_IFTYPE_MAX as the first invalid value
rather than as the last valid value. Update the code to allow a value of
exactly NL80211_IFTYPE_MAX.

The observed effect of this bug is that interfaces of type NaN (Wi-Fi
Aware) are being treated as invalid. This is mostly non-functional
because we don't support Wi-Fi Aware, but it does cause an error message
to be logged on Shill startup when a device supports NaN.

BUG=None
TEST=FEATURES=test emerge-brya shill
TEST=Before fix `restart shill` produces the line:
"ERR shill[976]: ERROR shill: [wifi_phy.cc(73)] Invalid iface type: 12"
After fix, no such line.

[1]http://cs/h/chromium/chromiumos/codesearch/+/main:src/third_party/kernel/upstream/include/uapi/linux/nl80211.h?l=3466

Change-Id: Ib6dc91fa1cda1f30b81f6836d2aa4674d91bd7c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5217772
Reviewed-by: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>